Split audio + mod commands into standalone files
